placed the search user logic in main route
