+++ app/templates/searchStudentPage.html
@@ -1,6 +1,5 @@
 {% extends "base.html"%}
 
-
 {% block scripts %}
   {{super()}}
 <script type="module" src="/static/js/searchStudent.js"></script>
@@ -8,14 +7,10 @@
 
 {% block styles %}
   {{super()}}
-
 {% endblock %}
 
-
 {% block app_content %}
-
 <form id= "searchStudent" action="volunteerProfile" method="post">
-
   <h1 align="center">Student Search Page</h1><br>
   <div class="col-sm-6 col-sm-offset-3 m-auto">
     <div class="row">
@@ -30,5 +25,4 @@
     </div>
   </div>
 </form>
-
 {% endblock %}
